Move account silence-related methods to concern (#28866)
